feat: Add Vehicle registration
It was configured an jwt token system that uses a pki system that is
generated, to sign and encrypt the tokens to make actions on the server.
Furthermore, it is created and stored a secret key used to craft an MAC
to verify with the license plate the user identity.